% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alterryx.R
\docType{package}
\name{alterryx}
\alias{alterryx}
\alias{alterryx-package}
\title{alterryx: an api client for the alteryx gallery}
\description{
alterryx provides functions to access each of the alteryx gallery api
endpoints.
}
\details{
For more information about the Alteryx Gallery API, see the documentation

To learn more about alterryx, start with the vignettes:
\code{browseVignettes(package = "alterryx")}
}

