% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx_app.R
\name{as.alteryx_app}
\alias{as.alteryx_app}
\alias{is.alteryx_app}
\title{Alteryx Applications}
\usage{
as.alteryx_app(x)

is.alteryx_app(object)
}
\arguments{
\item{x}{A \code{list} of values pertaining to an Alteryx app returned from
the Alteryx Gallery API}

\item{object}{An R object}
}
\description{
Handling of application resources returned by the Alteryx
Gallery API. For more information, see \code{\link{get_app}}
}
