% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx_app_admin.R
\name{as.alteryx_app_admin}
\alias{as.alteryx_app_admin}
\alias{is.alteryx_app_admin}
\title{Alteryx Applications (Admin)}
\usage{
as.alteryx_app_admin(x)

is.alteryx_app_admin(object)
}
\arguments{
\item{x}{A \code{list} of values pertaining to an Alteryx app returned from
the Alteryx Gallery Admin API}

\item{object}{An R object}
}
\description{
Handling of application resources returned by the Alteryx
Gallery Admin API. For more information, see \code{\link{get_app_admin}}
}
