% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx_job.R
\name{as.alteryx_job}
\alias{as.alteryx_job}
\alias{is.alteryx_job}
\title{Alteryx Jobs}
\usage{
as.alteryx_job(x, parent)

is.alteryx_job(object)
}
\arguments{
\item{x}{A \code{list} of values pertaining to an Alteryx job returned from
the Alteryx Gallery API}

\item{parent}{An \code{alteryx_app} or \code{alteryx_job} containing the id
and name information of the parent application for a job.}

\item{object}{An R object}
}
\description{
Handling of job information returned by the Alteryx
Gallery API. For more information, see \code{\link{get_job}}.
}
