% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{app_jobs}
\alias{app_jobs}
\alias{get_app_jobs}
\alias{get_job}
\title{Get App Jobs}
\usage{
get_app_jobs(app, request_params = list(),
  gallery = getOption("alteryx_gallery"))

get_job(job, gallery = getOption("alteryx_gallery"))
}
\arguments{
\item{app}{An Alteryx app returned from \code{\link{get_app}}}

\item{request_params}{List of request parameters. For more information on
parameters, visit the Alteryx Gallery API documentation
\url{https://gallery.alteryx.com/api-docs/}}

\item{gallery}{URL for your Alteryx Gallery}

\item{job}{An Alteryx job returned from \code{get_app_jobs} or
\code{\link{post_app_job}}}
}
\description{
A job is a single run of an app. To queue a job for an app use
\code{\link{queue_app}}. Once queued, there are two ways to poll the gallery
for an update on the job status: \code{get_app_jobs} or \code{get_job}.

\code{get_app_jobs} will return all jobs for a given app.
\code{get_job} will return a single job
}
\details{
Get the status of jobs running in Alteryx Gallery
}
\section{See Also}{

Once a job is complete, use \code{\link{download_job_output}} to retrieve
the results.
}
\examples{
\dontrun{
# get the five most recently queued jobs for an app
request_params <- list(
  sortField = "createdate",
  limit = "5"
)

get_app_jobs(app, request_params)
}
}

