% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{build_base_string}
\alias{build_base_string}
\title{Create the base string to be signed}
\usage{
build_base_string(request_method, base_url, normalized_request_params)
}
\arguments{
\item{request_method}{Character vector containing an HTTP request verb}

\item{base_url}{Base URL created with \code{build_base_url}}

\item{normalized_request_params}{Normalized request parameters created with
\code{normalize_request_params}}
}
\description{
Create the base string to be signed
}
