% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{build_base_url}
\alias{build_base_url}
\title{Build Base Url}
\usage{
build_base_url(gallery, endpoint)
}
\arguments{
\item{gallery}{URL of an Alteryx Gallery}

\item{endpoint}{API endpoint beginning and ending with "/"}
}
\description{
Create and percent encode the base URL, part of the base string to be signed
}
