% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/request.R
\name{build_request_body}
\alias{build_request_body}
\title{Build Request Body}
\usage{
build_request_body(name_value, ...)

build_request_body(name_value, ...)
}
\arguments{
\item{name_value}{\code{list} containing an app question \code{name} and
\code{value} pair}

\item{...}{Additional \code{name_value} pairs}

\item{name_value}{\code{list} containing an app question \code{name} and
\code{value} pair}

\item{...}{Additional \code{name_value} pairs}
}
\description{
Build the JSON of name value pairs corresponding to app questions for

Build the JSON of name value pairs corresponding to app questions for
}
\examples{
first_question <- list(name = "a", value = "1")
second_question <- list(name = "b", value = "2")
build_request_body(first_question, second_question)
first_question <- list(name = "a", value = "1")
second_question <- list(name = "b", value = "2")
build_request_body(first_question, second_question)
}

