% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{build_signature}
\alias{build_signature}
\title{Make Signature}
\usage{
build_signature(gallery, endpoint, request_method, required_headers,
  request_params)
}
\arguments{
\item{gallery}{URL of an Alteryx Gallery}

\item{endpoint}{API endpoint beginning and ending with "/"}

\item{request_method}{HTTP request verb}

\item{required_headers}{Required headers created with
\code{build_required_headers}}

\item{request_params}{List of request parameters}
}
\description{
Make Signature
}
