% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_class}
\alias{check_class}
\title{Error handling for incorrect object inputs to api.R calls}
\usage{
check_class(object, type)
}
\arguments{

}
\description{
Error handling for incorrect object inputs to api.R calls
}
