% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Status Code Check}
\usage{
check_status(response)
}
\arguments{
\item{response}{Response generated by \code{httr::GET()}}
}
\description{
Error handling for non-200 status codes
}

