% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_app}
\alias{download_app}
\title{Download App}
\usage{
download_app(app, destfile, gallery = getOption("alteryx_gallery"))
}
\arguments{
\item{app}{A single \code{alteryx_app} returned from \code{get_app}}

\item{destfile}{A character string with the directory and name of where the
downloaded file is to be saved.}

\item{gallery}{URL for your Alteryx Gallery}
}
\description{
Download an app as a .yxzp file
}
