% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_api.R
\name{download_app_admin}
\alias{download_app_admin}
\title{Get Applications (Admin)}
\usage{
download_app_admin(app, gallery = get_gallery())
}
\arguments{
\item{app}{A single \code{alteryx_app} returned from \code{get_app}}

\item{gallery}{URL for your Alteryx Gallery}
}
\value{
Raw binary for the resource .yxwz
}
\description{
Just like \code{download_app}, except when you're using admin keys. If your
not doing migration or management, you probably don't need this function.
}
\details{
This requires a separate function because more information is returned for
a resource by the admin API than the generic API.
}
