% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_answer}
\alias{empty_answer}
\title{Empty Answer}
\usage{
empty_answer()
}
\description{
Utility function used to queue a job for an app that has no questions
}
\examples{
\dontrun{
job <- queue_job(app, answer = empty_answer())
}

}
