% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_api.R
\name{get_app_admin}
\alias{get_app_admin}
\title{Get Applications (Admin)}
\usage{
get_app_admin(request_params = list(), gallery = get_gallery())
}
\arguments{
\item{request_params}{List of app search parameters. For more information on
parameters, visit the Alteryx Gallery API documentation
\url{https://gallery.alteryx.com/api-docs/} and see the parameters under the
'Find workflows in a subscription' section.}

\item{gallery}{URL for your Alteryx Gallery}
}
\description{
Just like \code{get_app}, except when you're using admin keys. If your not
doing migration or management, you probably don't need this function.
}
\details{
This requires a separate function because more information is returned for
a resource by the admin API than the generic API.
}
