% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_app_jobs}
\alias{get_app_jobs}
\title{Get App Jobs}
\usage{
get_app_jobs(gallery, app_id, request_params = list())
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{app_id}{ID of the app for which you want to retrieve the questions}

\item{request_params}{List of request parameters}
}
\description{
Returns the jobs for the given Alteryx Analtytics App. Only app workflows
can be used.
}

