% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_app_questions}
\alias{get_app_questions}
\title{Get App Questions}
\usage{
get_app_questions(gallery, app)
}
\arguments{
\item{app_id}{ID of the app for which you want to retrieve the questions}
}
\description{
Get the questions for the given Alteryx Analtytic App. Only app workflows
can be used.
}

