% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_apps}
\alias{get_apps}
\title{Get Subscriptions}
\usage{
get_apps(gallery, request_params = list())
}
\description{
Get the workflows in a subscription. Subscription is tied to API key. You
cannot request workflows for any other subscription without that
subscription's key.
}

