% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_job}
\alias{get_job}
\title{Get Job}
\usage{
get_job(gallery, job, request_params = list())
}
\arguments{
\item{job_id}{The ID of the job to retrieve}
}
\description{
Retrieves the job and its current state. Only app workflows can be used.
}

