% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_job_log}
\alias{get_job_log}
\title{Job Log}
\usage{
get_job_log(job)
}
\arguments{
\item{job}{An completed Alteryx job returned from \code{get_job}}
}
\description{
Retrieve the log messages for a "Completed" job
}
