% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_migration_form}
\alias{get_migration_form}
\title{Get Migration Default Form}
\usage{
get_migration_form(app)
}
\arguments{
\item{app}{A single \code{alteryx_app} returned from \code{get_app}}
}
\description{
Utility function to extract app info to populate migration form defaults
}
