% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_priority}
\alias{get_priority}
\title{Get Priority}
\usage{
get_priority(priority)
}
\arguments{
\item{priority}{A value of 'low', 'medium', 'high', or 'critical' to be
translated into the appropriate integer value for job priority}
}
\description{
Translate Gallery priority text to appropriate value
}
