% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_request_url}
\alias{get_request_url}
\title{Get Request URL}
\usage{
get_request_url(gallery, endpoint, request_method, request_params)
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_method}{HTTP request verb}

\item{request_params}{List of request parameters}
}
\description{
This provides a manual way to access Alteryx Gallery endpoints
but in general the specific endpoint functions should be used
}

