% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_subscriptions}
\alias{get_subscriptions}
\title{Get Subscriptions}
\usage{
get_subscriptions(gallery, request_params = list())
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{request_params}{List of request parameters}
}
\description{
Get the workflows in a subscription. Subscription is tied to API key. You
cannot request workflows for any other subscription without that
subscription's key.
}

