% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{make_base_string}
\alias{make_base_string}
\title{Make Base String}
\usage{
make_base_string(request_method, base_url, normalized_request_params)
}
\arguments{
\item{request_method}{HTTP request verb}

\item{base_url}{Base url created with \code{make_base_url}}

\item{normalized_request_params}{Normalized request parameters created with
\code{normalize_request_params}}
}
\description{
Create the base string to be signed
}

