% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{make_base_url}
\alias{make_base_url}
\title{Make Base Url}
\usage{
make_base_url(gallery, endpoint)
}
\arguments{
\item{gallery}{The url an Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}
}
\description{
Create and url encode the base url
}

