% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{make_required_headers}
\alias{make_required_headers}
\title{Generate Required Headers}
\usage{
make_required_headers()
}
\description{
Alteryx Gallery requests require headers consumer_key, signature_method,
timestamp, nonce, and version.
}

