% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{make_signature}
\alias{make_signature}
\title{Make Signature}
\usage{
make_signature(gallery, endpoint, request_method, required_headers,
  request_params)
}
\arguments{
\item{gallery}{The url an Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_method}{HTTP request verb}

\item{required_headers}{Required headers created with
\code{make_required_headers}}

\item{request_params}{List of request parameters
such as \code{list(values = json_body)}}
}
\description{
Make Signature
}

