% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{manage_job_output}
\alias{download_job_output}
\alias{job_output}
\alias{manage_job_output}
\title{Manage Job Output}
\usage{
job_output(job)

download_job_output(job, filename = names(job_output(job)),
  download_directory = tempdir(), request_params = list(),
  gallery = getOption("alteryx_gallery"))
}
\arguments{
\item{job}{An completed Alteryx job returned from \code{get_app_jobs} or
\code{\link{post_app_job}}}

\item{filename}{A character vector of filenames for the output}

\item{download_directory}{Directory in which to write the outputs}

\item{request_params}{List of request parameters. For more information on
parameters, visit the Alteryx Gallery API documentation
\url{https://gallery.alteryx.com/api-docs/}}

\item{gallery}{URL for your Alteryx Gallery}
}
\description{
If an Alteryx app includes one or more output tools inside the
application, the output will be available for download once the job is
complete.

Use \code{job_output} to get information about the job outputs.
Use \code{download_job_output} to download the job outputs and save them on
your local machine.
}
\details{
Get the default names and download outputs from an Alteryx job.
}

