% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{parse_request_response}
\alias{parse_request_response}
\title{Parse Request Response}
\usage{
parse_request_response(response, as = "text", remove_bom = TRUE,
  parse_JSON = TRUE)
}
\arguments{
\item{response}{Response generated by a request such as \code{httr::GET}
or \code{httr::POST}}

\item{as}{Passed to \code{httr::content(as = as)}}

\item{remove_bom}{Remove BOM from response content?}

\item{parse_JSON}{Parse response content as JSON?}
}
\description{
Parse Request Response
}
