% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_app_job}
\alias{post_app_job}
\title{Queue Job}
\usage{
post_app_job(gallery, app, request_body)
}
\arguments{
\item{request_body}{JSON body of request containing values for app questions
built using \code{build_request_body}}

\item{app_id}{ID of an app for which to queue a job}
}
\value{
Returns the ID of the job. Use the status request to get the results
of the job.
}
\description{
Queue a job for an application
}

