% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{remove_bom}
\alias{remove_bom}
\title{Remove Byte Order Mark}
\usage{
remove_bom(content)
}
\arguments{
\item{content}{Content from response generated by \code{httr::content()}}
}
\description{
Janky way to remove BOM from response content
}

