% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gallery.R
\name{run_gallery}
\alias{run_gallery}
\title{Launch Gallery Application}
\usage{
run_gallery()
}
\description{
This package includes an experimental Shiny application that uses the
\code{alterryx} functions to mimic Alteryx Gallery functionality.
The goal of this is to understand how to use the API to extend Gallery
functionality. Big thanks to Dean Attali for his primer on including
Shiny apps in an R package.
}
