% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{sign_base_string}
\alias{sign_base_string}
\title{Sign Base String}
\usage{
sign_base_string(base_string, alteryx_secret_key)
}
\arguments{
\item{base_string}{Base string created with \code{build_base_string}}

\item{alteryx_secret_key}{Admin API secret key for target environment}
}
\description{
Sign Base String
}
