% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_get_request}
\alias{submit_get_request}
\title{Submit Get Request}
\usage{
submit_get_request(gallery, endpoint, request_params)
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_params}{List of request parameters}
}
\description{
Submit Get Request
}

