% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{submit_put_request}
\alias{submit_put_request}
\title{Submit PUT Request}
\usage{
submit_put_request(gallery, endpoint, request_params,
  alteryx_api_key = getOption("alteryx_api_key"),
  alteryx_secret_key = getOption("alteryx_secret_key"), ...)
}
\arguments{
\item{gallery}{URL for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_params}{List of request parameters}

\item{alteryx_api_key}{Admin API key for target environment}

\item{alteryx_secret_key}{Admin API secret key for target environment}

\item{...}{Additional arguments passed to \code{parse_request_response}}
}
\description{
Submit PUT Request
}
