% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_request}
\alias{submit_request}
\title{Submit Request}
\usage{
submit_request(gallery, endpoint, request_method, request_params)
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_method}{HTTP request verb}

\item{request_params}{List of request parameters}
}
\description{
Submit Request
}

