% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{build_request_body}
\alias{build_request_body}
\title{Build Request Body}
\usage{
build_request_body(name_value, ...)
}
\arguments{
\item{name_value}{a \code{list} containing the app question \code{name} and
\code{value} pairs}

\item{...}{additional \code{name_value} pairs}
}
\description{
Build the JSON of name value pairs corresponding to app questions
}
\examples{
first_question <- list(name = "a", value = "1")
second_question <- list(name = "b", value = "2")
build_request_body(first_question, second_question)
}

