% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{build_request_url}
\alias{build_request_url}
\title{Build Request URL}
\usage{
build_request_url(gallery, endpoint, request_method, request_params)
}
\arguments{
\item{gallery}{URL for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_method}{HTTP request verb}

\item{request_params}{List of request parameters}
}
\description{
This provides a manual way to build a request URL but in general this
should be handled by the specific endpoint functions
}

