% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{check_status}
\alias{check_status}
\title{Check Response Status}
\usage{
check_status(response)
}
\arguments{
\item{response}{Response generated by a request such as \code{httr::GET}
or \code{httr::POST}}
}
\description{
Error handling for non-200 status codes
}

