% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_list}
\alias{encode_list}
\title{URL Encode List}
\usage{
encode_list(l, reserved = TRUE)
}
\arguments{
\item{l}{a list}

\item{reserved}{logical: should ‘reserved’ characters be encoded?}
}
\description{
URL encode all values in a list using \code{utils::URLencode()}
}

