% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_app_questions}
\alias{get_app_questions}
\title{Get App Questions}
\usage{
get_app_questions(app, gallery = getOption("alteryx_gallery"))
}
\arguments{
\item{app}{A single \code{alteryx_app} returned from \code{get_app}}

\item{gallery}{URL for your Alteryx Gallery}
}
\description{
Get the questions for the given Alteryx Analtytic App. Only app
workflows can be used.
}
\details{
Most Alteryx apps have questions, user input that defines how
the application should run. The answers to these questions need to be sent
as \code{answers} when queueing an application on Gallery with
\code{queue_job}. \code{get_app_questions} returns the names,
types, and default values of the questions for an app.
}
\section{WARNING}{

Trees, Maps, and File Browse are unsupported question types. This is a
limitation of Alteryx Gallery and not this API client.
}

