% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_info}
\alias{get_info}
\title{Get Resource Information}
\usage{
get_info(resource)
}
\arguments{
\item{resource}{An Alteryx \code{app} or \code{job}}
}
\description{
Retrieves all metadata about an Alteryx resource including but not limited
to upload date, version, and author
}
