% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_job_output}
\alias{get_job_output}
\title{Get Job Output}
\usage{
get_job_output(gallery, job, request_params = list())
}
\arguments{
\item{output_id}{The ID for the output you want to retrieve.}
}
\description{
Returns output for a given job. Only app workflows can be used.
}

