% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{normalize_request_params}
\alias{normalize_request_params}
\title{Normalize Request Parameters}
\usage{
normalize_request_params(required_headers, request_params)
}
\arguments{
\item{required_headers}{Required headers created with
\code{build_required_headers}}

\item{request_params}{List of request parameters}
}
\description{
Combine, sort, and percent encode headers and param list, part of the base
string to be signed
}
