% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{sign_base_string}
\alias{sign_base_string}
\title{Sign Base String}
\usage{
sign_base_string(base_string)
}
\arguments{
\item{base_string}{base string created with \code{make_base_string}}
}
\description{
Sign Base String
}

