% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{submit_get_request}
\alias{submit_get_request}
\title{Submit GET Request}
\usage{
submit_get_request(gallery, endpoint, request_params, ...)
}
\arguments{
\item{gallery}{URL for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_params}{List of request parameters}

\item{...}{Additional arguments passed to \code{parse_request_response}}
}
\description{
Submit GET Request
}
