% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{submit_post_request}
\alias{submit_post_request}
\title{Submit POST Request}
\usage{
submit_post_request(gallery, endpoint, request_params, request_body, ...)
}
\arguments{
\item{gallery}{URL for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_params}{List of request parameters}

\item{request_body}{JSON body of request containing values for app questions
built using \code{build_request_body}}

\item{...}{Additional arguments passed to \code{httr::POST()} such as
\code{verbose}}
}
\description{
This provides a manual way to submit a POST request but in general this
should be handled by the specific endpoint functions
}

