% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_post_request}
\alias{submit_post_request}
\title{Submit Post Request}
\usage{
submit_post_request(gallery, endpoint, request_body)
}
\arguments{
\item{gallery}{The url for your Alteryx Gallery}

\item{endpoint}{The api endpoint beginning and ending with "/"}

\item{request_body}{JSON body of request containing values for app questions
built using \code{build_request_body}}
}
\description{
Submit Post Request
}

